+++ src/main/java/com/ruoyi/project/vocsale/domain/dto/ListVocSaleResponseDto.java
@@ -5,8 +5,6 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 /**
  * Created by Fyc on 2021-8-2.
  * 获取销售信息列表
@@ -22,10 +20,10 @@ public class ListVocSaleResponseDto extends BaseResponseDto
     private String saleType;
 
     @ApiModelProperty(value = "销售总量")
-    private int totalAmount;
+    private Integer totalAmount;
 
     @ApiModelProperty(value = "销售总价")
-    private BigDecimal totalPrice;
+    private Double totalPrice;
 
     @ApiModelProperty(value = "销售状态")
     private String saleStatus;
